# services/core/docs/swagger.yaml
basePath: /api/v1
definitions:
  ai.GeneratedPlan:
    properties:
      description:
        type: string
      tasks:
        items:
          $ref: '#/definitions/ai.GeneratedTask'
        type: array
      title:
        type: string
    type: object
  ai.GeneratedTask:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  dto.GeneratePlanRequest:
    properties:
      description:
        type: string
      position:
        type: string
      topic:
        type: string
    required:
    - topic
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  dto.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RefreshResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Mandzhonson
  description: REST API для SkillTracker.
  title: SkillTracker API
  version: "1.0"
paths:
...
  /ai/generate-plan:
    post:
      consumes:
      - application/json
      description: Генерирует план обучения на основе темы и описания
      parameters:
      - description: Данные для генерации плана
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GeneratePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai.GeneratedPlan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Сгенерировать план с помощью ИИ
      tags:
      - AI
  /auth/login:
    post:
      consumes:
      - application/json
      description: Авторизация пользователя по email и паролю
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Авторизация
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Инвалидирует refresh токен
      parameters:
      - description: Refresh токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Выход из системы
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Обновляет access и refresh токены
      parameters:
      - description: Refresh токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RefreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Обновить токены
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"